Add event listener that plays a round and updates score
Add event listener for each button which stores target id as
in a variable called choice. Then sends choice to function onClick.

onClick function takes choice and calls play round with choice as the
parameter for player selection.

add condition that onCLick is called only when playerScore and computer
Score dont equal 5.

Add a div called GG to display game over message when player or
computer reach a score of 5.